Sample game data
I came up with a sample file specifying game content. I'll work on the actual parser later. Things do *not* have to look this way if we'd rather do things differently, but this looks like a good way to start us all thinking about what we want the game to be able to do, and how we might be able to separate content in an extensible way.